pyproject.toml
[project]
name = "pandas-lib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pymongo>=4.16.0",
    "sqlalchemy>=2.0.48",
]
